requirements.txt
Flask==3.1.0
gunicorn==23.0.0
anthropic==0.49.0
google-api-python-client==2.163.0
google-auth-oauthlib==1.2.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
